// gdr-web-version-mdp/gdr-web-version-mdp/src/app/@data/model/organo.ts
export class Organo {
    organigramaId?: number;
    entidadId: number;
    areaId: number;
    paisId?: number;
    sedeId: number;
    padreOrganigramaId?: number;
    puesto: string;
    orden: number;
    nivel: number;
    tipoOrganoUoId: number;
    naturalezaOrgano: number;
    nivelGobiernoId: number;
    descripcion: string;
    descripcionCorta?: string;
    sigla: string;
    personaResponsableId: string;
    telefonoId?: number;
    correoId?: number;
    descripcionTipoOrg: string;
    estado?: string;
    nombres: string;
    apellidoPaterno: string;
    apellidoMaterno?: string;
    telefono: string;
    correo: string;
    tipoDocumento: number;
    nroDocumento: string;
    descripOrganoPadre: string;
    nombrePais?: string;
    desNivel: string;
    desNaturaleza: string;
    estadoRegistro?: string;
    destipoDocumento?: string;
    tipo?: number;

}



/* export class Organo {
    estadoRegistro?: string;
    usuarioCreacion?: string;
    fechaCreacio?: string;
    usuarioModificacion?: string;
    fechaModificacion?: string;
    puestoId?: number;
    organigramaId: number;
    entidadId: number;
    areaId: number;
    sedeId: number;
    padreOrganigramaId?: number;
    puesto: string;
    orden: number;
    nivel: number;
    tipoOrganoUoId: number;
    naturalezaOrgano: number;
    nivelGobiernoId: number;
    descripcion: string;
    descripcionCorta?: string;
    sigla: string;
    personaResponsableId: string;
    telefono: string;
    telefonoId?: number;
    correo: string;
    correoId?: number;
    apellidoMaterno?: string;
    apellidoPaterno: string;
    desNaturaleza: string;
    desNivel: string;
    descripOrganoPadre: string;
    descripcionTipoOrg: string;
    estado?: string;
    nombrePais?: string;
    nombres: string;
    nroDocumento: string;
    tipoDocumento: number;
    settings: any;
    tipo?: number;
}
  */